Remove migration for Products-Brands relationship
The migration class `_1_update_brand_table_relationship` has been removed. This class added a `BrandId` column to the `Products` table, created an index on this column, and established a foreign key constraint linking `BrandId` in `Products` to `BrandId` in `Brands`. The `Down` method, which would reverse these changes, is also removed.